+++ src/pages/programsWip/form/sectionForm/EntrollmentSectionForm.tsx
@@ -3,8 +3,8 @@ import { useQuery } from '@tanstack/react-query'
 import arrayMutators from 'final-form-arrays'
 import React, { useCallback, useMemo } from 'react'
 import { useParams } from 'react-router-dom'
-import { FormBase, FormBaseProps } from '../../../../../components'
-import { LoadingSpinner } from '../../../../../components/loading/LoadingSpinner'
+import { FormBase, FormBaseProps } from '../../../../components'
+import { LoadingSpinner } from '../../../../components/loading/LoadingSpinner'
 import {
     DEFAULT_FIELD_FILTERS,
     SchemaName,
@@ -14,102 +14,86 @@ import {
     createFormError,
     createJsonPatchOperations,
     useCreateModel,
-} from '../../../../../lib'
+} from '../../../../lib'
 import {
     DisplayableModel,
     PickWithFieldFilters,
     Section,
-} from '../../../../../types/models'
-import { DataSetSectionFormContents } from './DataSetSectionFormContents'
+} from '../../../../types/models'
+import { EnrollmentSectionFormContents } from './EnrollmentSectionFormContents'
 import { initialSectionValues } from './sectionFormSchema'
 
 export const fieldFilters = [
     ...DEFAULT_FIELD_FILTERS,
     'name',
     'description',
-    'code',
-    'indicators[id,displayName]',
-    'showRowTotals',
-    'showColumnTotals',
-    'disableDataElementAutoGroup',
-    'displayOptions',
-    'dataElements[id,displayName,categoryCombo[id]]',
-    'greyedFields[dataElement, categoryOptionCombo]',
+    'renderType[MOBILE[type],DESKTOP[type]]',
+    'trackedEntityAttributes[id,displayName]',
 ] as const
 
-const dataSetSectionSchemaSection = {
-    name: SchemaName.section,
-    namePlural: 'sections',
+export const enrollmentSectionSchemaSection = {
+    name: SchemaName.programSection,
+    namePlural: 'programSections',
     title: i18n.t('Section'),
     titlePlural: i18n.t('Sections'),
-    parentSectionKey: 'dataSet',
+    parentSectionKey: 'programsAndTracker',
 } satisfies SchemaSection
 
 export type SectionFormValues = PickWithFieldFilters<
     Section,
     typeof fieldFilters
 > & {
-    dataSet: { id: string }
-    displayOptions?: string
+    program: { id: string }
 }
+
 type PartialSectionFormValues = Partial<SectionFormValues>
 type SubmittedSectionFormValues = PartialSectionFormValues & DisplayableModel
 
-export type DataSetSectionFormProps = {
-    dataSetSection?: PartialSectionFormValues
+export type EnrollmentSectionFormProps = {
+    enrollmentSection?: PartialSectionFormValues
     onCancel?: () => void
+    sortOrder?: number
 } & Pick<FormBaseProps<PartialSectionFormValues>, 'onSubmit'>
 
-export const DataSetSectionForm = ({
-    dataSetSection,
+export const EnrollmentSectionForm = ({
+    enrollmentSection,
     onSubmit,
     onCancel,
-}: DataSetSectionFormProps) => {
-    const dataSetId = useParams().id as string
+    sortOrder,
+}: EnrollmentSectionFormProps) => {
+    const programId = useParams().id as string
+
     const initialValues: PartialSectionFormValues | undefined = useMemo(() => {
-        if (dataSetSection) {
-            return {
-                ...dataSetSection,
-                displayOptions:
-                    dataSetSection?.displayOptions &&
-                    JSON.parse(dataSetSection?.displayOptions),
-            }
+        if (enrollmentSection) {
+            return enrollmentSection
         }
-        return {
-            ...initialSectionValues,
-            displayOptions:
-                initialSectionValues?.displayOptions &&
-                JSON.parse(initialSectionValues?.displayOptions),
-        }
-    }, [dataSetSection])
+        return { ...initialSectionValues, sortOrder }
+    }, [enrollmentSection, sortOrder])
 
     const valueFormatter = useCallback(
         (values: PartialSectionFormValues) => {
             return {
                 ...values,
-                dataSet: { id: dataSetId },
-                displayOptions:
-                    values.displayOptions &&
-                    JSON.stringify(values.displayOptions),
+                program: { id: programId },
             }
         },
-        [dataSetId]
+        [programId]
     )
     return (
         <FormBase
-            initialValues={{ ...initialValues, dataSet: { id: dataSetId } }}
+            initialValues={{ ...initialValues, program: { id: programId } }}
             onSubmit={onSubmit}
             valueFormatter={valueFormatter}
             includeAttributes={false}
             mutators={{ ...arrayMutators }}
         >
-            <DataSetSectionFormContents onCancel={onCancel} />
+            <EnrollmentSectionFormContents onCancel={onCancel} />
         </FormBase>
     )
 }
 
 type OnDataSetFormSubmit = FormBaseProps<PartialSectionFormValues>['onSubmit']
-export const EditDataSetSectionForm = ({
+export const EditEnrollmentSectionForm = ({
     section,
     onCancel,
     onSubmitted,
@@ -120,7 +104,7 @@ export const EditDataSetSectionForm = ({
 }) => {
     const handlePatch = usePatchModel(
         section.id,
-        dataSetSectionSchemaSection.namePlural
+        enrollmentSectionSchemaSection.namePlural
     )
 
     const onFormSubmit: OnDataSetFormSubmit = async (values, form) => {
@@ -152,7 +136,7 @@ export const EditDataSetSectionForm = ({
         queryFn: queryFn<SectionFormValues>,
         queryKey: [
             {
-                resource: 'sections',
+                resource: 'programSections',
                 id: section.id,
                 params: {
                     fields: fieldFilters.concat(),
@@ -166,22 +150,26 @@ export const EditDataSetSectionForm = ({
     }
 
     return (
-        <DataSetSectionForm
-            dataSetSection={sectionValues.data}
+        <EnrollmentSectionForm
+            enrollmentSection={sectionValues.data}
             onSubmit={onFormSubmit}
             onCancel={onCancel}
         />
     )
 }
 
-export const NewDataSetSectionForm = ({
+export const NewEnrollmentSectionForm = ({
     onCancel,
     onSubmitted,
+    sortOrder,
 }: {
     onCancel: () => void
     onSubmitted: (values: SubmittedSectionFormValues) => void
+    sortOrder: number
 }) => {
-    const handleCreate = useCreateModel(dataSetSectionSchemaSection.namePlural)
+    const handleCreate = useCreateModel(
+        enrollmentSectionSchemaSection.namePlural
+    )
 
     const onFormSubmit: OnDataSetFormSubmit = async (values) => {
         const res = await handleCreate(values)
@@ -199,34 +187,41 @@ export const NewDataSetSectionForm = ({
     }
 
     return (
-        <DataSetSectionForm
-            dataSetSection={undefined}
+        <EnrollmentSectionForm
+            enrollmentSection={undefined}
             onSubmit={onFormSubmit}
             onCancel={onCancel}
+            sortOrder={sortOrder}
         />
     )
 }
 
-export const EditOrNewDataSetSectionForm = ({
+export const EditOrNewEnrollmentSectionForm = ({
     section,
     onCancel,
     onSubmitted: onSubmit,
+    sectionsLength,
 }: {
     section: DisplayableModel | null | undefined
     onCancel: () => void
     onSubmitted: (values: SubmittedSectionFormValues) => void
+    sectionsLength: number
 }) => {
     if (section === undefined) {
         return
     }
     if (section === null) {
         return (
-            <NewDataSetSectionForm onSubmitted={onSubmit} onCancel={onCancel} />
+            <NewEnrollmentSectionForm
+                onSubmitted={onSubmit}
+                onCancel={onCancel}
+                sortOrder={sectionsLength}
+            />
         )
     }
 
     return (
-        <EditDataSetSectionForm
+        <EditEnrollmentSectionForm
             section={section}
             onCancel={onCancel}
             onSubmitted={onSubmit}